Get rid of route manager
